feat: optimize axios code
